Change code to adhere to code quality in Week 5's topics, and fix a critical bug regarding find function in Angela.
In Week 5's code quality topics, it is recommended to add a default
branch for case statements, handlePrintTask and
handleTaskModification fail this code quality check.

For the find function bug, I have taken notice that Angela will
throw IndexOutOfBound exception when entering the command "find "
(with a space). This has been fixed.