add Assertions to Command.java and add banner and icon to GUI of Angela
Assertions are added to handle command functions to terminate the
program when a command is erroneously sent to the function. This
add another layer of guard from unexpected behaviours.

Angela GUI will now be displayed as Angela in the window banner,
and Lobotomy Corp Icon will be displayed as window Icon.